FEAT+FIX: Greeting and minor fixes
* FEAT: Greeting feature (like in text-generation webui)
* FIX: !ai and !chat prefixes removed from the user's message
**Backstory of the PR:**
One day... I was just sitting and scrolling through the text generation webui's API, looking for the settings that could improve the chatbot. My eyes catched the "greeting" string, and I tried to use it. I added the setting into CUSTOM_MODEL_SETTINGS and reloaded the program, after typing "!ai Hello World!" in the chat I discovered that this setting does not work, or at least not the way I think it would. Filled with sadge I decided to code a simple solution myself.